Extract templating from provisioner/node_pools
